Use explicit typecast to size_t
maxMemoryUsageBytes is size_t type which may not match long long value
internalDocumentSourceGroupMaxMemoryBytes.load() returns, so typecast it
to avoid narrowing warning [-Wc++11-narrowing] from clang

+++ src/mongo/db/pipeline/document_source_group.cpp
@@ -399,7 +399,7 @@ DocumentSourceGroup::DocumentSourceGroup(const intrusive_ptr<ExpressionContext>&
       _doingMerge(false),
       _memoryTracker{expCtx->allowDiskUse && !expCtx->inMongos,
                      maxMemoryUsageBytes ? *maxMemoryUsageBytes
-                                         : internalDocumentSourceGroupMaxMemoryBytes.load()},
+                                         : (size_t)internalDocumentSourceGroupMaxMemoryBytes.load()},
       _initialized(false),
       _groups(expCtx->getValueComparator().makeUnorderedValueMap<Accumulators>()),
       _spilled(false) {
